pyproject.toml
[project]
name = "anomalies-pieces-indus"
version = "0.1.0"
description = "Détection d'anomalies sur pièces industrielles"
requires-python = ">=3.12"
dependencies = [
    # Anomaly detection
    "anomalib==2.3.1",
    "torch>=2.1",
    "torchvision>=0.16",
    "lightning>=2.1",
    "timm>=1.0",
    # Data & scientific
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "Pillow>=10.0",
    # Visualization
    "matplotlib>=3.7",
    "seaborn>=0.13",
    # Computer vision
    "opencv-python>=4.8",
]

[project.optional-dependencies]
app = [
    "streamlit>=1.30",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
